backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
motor==3.7.0
redis==5.2.1
python-dotenv==1.0.1
